Add script for changing talk states in PaperCall
Changing talks from one state to another is a huge hassle. PaperCall
doesn't have bulk-edit functionality, so you have to do it one-by-one.
Furthermore, it takes seconds for state changes to apply. Worse,
talks are paginated and PaperCall's search functionality is limited
to tags, so it can take a while to find the talk to transition before
waiting 30 years for the transition to complete.

This script solves that problem.

Use this script to change the status of one or more talks, like this:

```sh
$: ./change_talk_state.sh --event "Name of event" \
  --talks "TALK_1|TALK_2|TALK_3|..." \
  --state "accepted"
```

You'll first need to retrieve a valid session token from PaperCall. This
can be done from your browser. See `./change_talk_state.sh --help` for
more information.